git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11224 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/shared/TransformTierGlyph.java
@@ -115,7 +115,7 @@ public final class TransformTierGlyph extends TransformViewModeGlyph {
     LinearTransform view_transform = view.getTransform();
     double yscale = 0.0d;
     if ( 0.0d != getCoordBox().height ) {
-      yscale = (double)fixedPixHeight / getCoordBox().height;
+      yscale = fixedPixHeight / getCoordBox().height;
     }
 	if ( 0.0d != view_transform.getScaleY() ) {
 	  yscale = yscale / view_transform.getScaleY();
